Refactor: move 'print welcome message' from Biblioteca App to Printer
